#83 Fixed compilation warnings and redundancies

+++ examples/Example.LightBDD.NUnit2/Features/Calculator_feature.Steps.cs
@@ -15,19 +15,19 @@ namespace Example.LightBDD.NUnit2.Features
 
         private void Then_adding_X_to_Y_should_give_RESULT(int x, int y, int result)
         {
-            Assert.AreEqual(result, (int) _calculator.Add(x, y));
+            Assert.AreEqual(result, _calculator.Add(x, y));
         }
 
         private void Then_dividing_X_by_Y_should_give_RESULT(int x, int y, int result)
         {
-            Assert.AreEqual(result, (int) _calculator.Divide(x, y));
+            Assert.AreEqual(result, _calculator.Divide(x, y));
         }
 
         private void Then_multiplying_X_by_Y_should_give_RESULT(int x, int y, int result)
         {
             if (x < 0 || y < 0)
                 Assert.Inconclusive("Negative numbers are not supported yet");
-            Assert.AreEqual(result, (int) _calculator.Multiply(x, y));
+            Assert.AreEqual(result, _calculator.Multiply(x, y));
         }
     }
 }
